Fix: [star-stock] 일부 버그 수정
1) 별표 호버 시 포인터 적용

2) 관심 종목이 없을 경우 메시지 추가
