+++ assets/css/stock-detail.css
@@ -143,6 +143,6 @@
   height: 32px;
 }
 
-.star-stock:hover {
+.star-stock img:hover {
   cursor: pointer;
 }
