+++ assets/js/common.js
@@ -278,3 +278,22 @@ function kakaoLogin() {
 function naverLogin() {
   window.location.href = "/api/naver/login";
 }
+
+/* 시가 총액 한글 표현으로 계산하는 함수 */
+export const formatPrice = (number) => {
+  const trillion = Math.floor(number / 1e12);
+  const billion = Math.floor((number % 1e12) / 1e8);
+
+  const formattedNumber = [];
+
+  if (trillion > 0) {
+    formattedNumber.push(`${trillion.toLocaleString()}조`);
+  }
+
+  if (billion > 0) {
+    const billionString = billion.toLocaleString();
+    formattedNumber.push(`${billionString}억`);
+  }
+
+  return formattedNumber.join(' ');
+}
